Update README.md
added instructions for setting up the api endpoints.  apparently the ones i have setup, because the ollama endpoint is pointed at my local server, it locks the program and desktop up on a linux based system.  so im removing and just giving instructions on how to connect them in the api manager.