Assistant node and ExportWord updates
This update contains significant changes to the assistant node as well as its ability to interact with the ExportWord.py code when it comes to exporting exporting and formatting the document but also drastic improvements when it comes to its ability to export charts and diagrams based on the user request and given data. Made improvements with the installer so that it's a little cleaner and should work better on Windows systems. Still need to do testing on Linux and Mac systems.