updated client installer
This update involves adding installers that will install the necessary requirements to run the assistant client. They are self-contained and may ask you to just continue installing if you want it to do that. There is a BAT file for Windows systems, and a sh bash file for Linux and Mac systems.  This client is used in conjunction with the assistant node. When you set up an assistant you would set up the Inbox and outbox folders, then you can run this client as a standalone and point it to the same inbox and outbox directories. This client interface gives you more functionality than the node interface itself.