updated export word functionality
This update involves rewriting some of the export functionality as well as giving it the ability to save the individual graphic files that are required from the chart and diagram sources. Instead of only including them in the exported Word document. It will create a subfolder and save the graphics to those folders. This will be something that is mostly used when you're utilizing the assistant node.